requirements.txt
langgraph
langchain-core
langchain-openai
langchain-chroma 
langchain
numpy
ipython
python-dotenv
chromadb 
pandas
scikit-learn